Introduce Format Dictionary capabilities that handles settings
